pyproject.toml
[project]
name = "interview-preparation-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.13",
    "audio-recorder-streamlit>=0.0.10",
    "chromadb>=1.0.12",
    "duckduckgo-search>=8.0.4",
    "fastapi>=0.115.9",
    "google-adk>=1.3.0",
    "google-generativeai>=0.8.5",
    "gtts>=2.5.4",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "pydantic>=2.11.7",
    "pydub>=0.25.1",
    "pygame>=2.6.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.1",
    "requests>=2.32.4",
    "sentence-transformers>=4.1.0",
    "speechrecognition>=3.14.3",
    "streamlit>=1.45.1",
    "streamlit-audiorecorder>=0.0.6",
    "streamlit-chat>=0.1.1",
    "streamlit-javascript>=0.1.5",
    "streamlit-mic-recorder>=0.0.8",
    "uvicorn>=0.34.3",
    "websockets>=15.0.1",
]
